asl-classifier/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
matplotlib
python-multipart
opencv-python

